added separate component for category filter
