+++ src/pages/ProductListings.jsx
@@ -1,5 +1,5 @@
 import ProductGrid from "../components/ProductGrid";
-import Filters from "../components/Filters";
+import Filters from "../components/Filter/Filters";
 import "../styles/product-list.css";
 
 export default function ProductList() {
